Fixed cached expiration pruning
