+++ utils.py
@@ -1,9 +1,10 @@
 import sys
 import re
 import requests
+import os
 import dateutil.parser
 from colors import *
-from caching import *
+from caching import get_cached_user_list, get_cached_repo_list, cache_user_list, cache_repo_list, CACHING_ACTIVE
 
 
 MAX_PAGES = 5
@@ -190,19 +191,3 @@ def rate_limit_check(req):
             putln(yellow + bold, 'Rate Limit Warning! {} requests left. Appox. {} - {} invocations remaining'.format(
                 left, int(left / 6), int(left / 2)))
             clear()
-
-
-if os.environ.get('DEBUG', False):
-
-    import inspect
-
-    def line():
-        return inspect.getframeinfo(inspect.currentframe()).lineno
-
-    def debug(msg, color=black):
-        print("{3}DEBUG: [{0}:{1}] {2}{4}".format(
-            __name__, line(), msg, color, black))
-else:
-    def line(): return ''
-
-    def debug(*args, **kwargs): pass
